requirements.txt
# =============================================================================
# 基于多智能体的热点自适应保险营销内容生成系统 - Python依赖包
# =============================================================================

# -----------------------------------------------------------------------------
# 核心框架和基础库
# -----------------------------------------------------------------------------
# 异步编程和并发
asyncio>=3.4.3
threading2>=0.1.8

# 环境变量管理
python-dotenv>=1.0.0

# 日期时间处理
python-dateutil>=2.8.2

# JSON和数据处理
ujson>=5.7.0

# 配置文件处理
configparser>=5.3.0

# -----------------------------------------------------------------------------
# Elasticsearch相关
# -----------------------------------------------------------------------------
# Elasticsearch官方客户端
elasticsearch>=8.10.0

# 批量操作支持
elasticsearch-dsl>=8.9.0

# -----------------------------------------------------------------------------
# 大语言模型相关
# -----------------------------------------------------------------------------
# HTTP请求库
requests>=2.31.0
urllib3>=2.0.4

# JSON处理
jsonschema>=4.19.0

# -----------------------------------------------------------------------------
# 向量计算和机器学习
# -----------------------------------------------------------------------------
# 句子转换器（向量化）
sentence-transformers>=2.2.2

# PyTorch生态系统
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# 科学计算
numpy>=1.24.3
scipy>=1.11.1

# 数组操作
pandas>=2.0.3

# 机器学习工具
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# 网页爬取和内容提取
# -----------------------------------------------------------------------------
# HTML解析
beautifulsoup4>=4.12.2
lxml>=4.9.3

# HTTP客户端
httpx>=0.24.1

# 用户代理处理
fake-useragent>=1.4.0

# URL处理
urllib3>=2.0.4

# -----------------------------------------------------------------------------
# 文本处理和NLP
# -----------------------------------------------------------------------------
# 正则表达式增强
regex>=2023.6.3

# 文本清理
text-unidecode>=1.3

# 中文文本处理
jieba>=0.42.1

# 字符编码检测
chardet>=5.1.0

# -----------------------------------------------------------------------------
# 数据存储和序列化
# -----------------------------------------------------------------------------
# JSON处理
orjson>=3.9.2

# 数据序列化
pickle5>=0.0.12

# 缓存
cachetools>=5.3.1

# -----------------------------------------------------------------------------
# 网络和API
# -----------------------------------------------------------------------------
# 高级HTTP库
httpx>=0.24.1
aiohttp>=3.8.5

# 重试机制
tenacity>=8.2.2

# 超时处理
timeout-decorator>=0.5.0

# -----------------------------------------------------------------------------
# 日志和监控
# -----------------------------------------------------------------------------
# 日志库
loguru>=0.7.0

# 系统监控
psutil>=5.9.5

# 内存分析
memory-profiler>=0.61.0

# 性能分析
line-profiler>=4.0.3

# -----------------------------------------------------------------------------
# 开发和测试工具
# -----------------------------------------------------------------------------
# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.1

# 代码格式化
black>=23.7.0
isort>=5.12.0

# 代码检查
flake8>=6.0.0
pylint>=2.17.4

# 类型检查
mypy>=1.4.1

# -----------------------------------------------------------------------------
# 安全和加密
# -----------------------------------------------------------------------------
# 加密库
cryptography>=41.0.2

# 哈希函数
hashlib2>=1.0.1

# -----------------------------------------------------------------------------
# 系统和进程管理
# -----------------------------------------------------------------------------
# 进程管理
supervisor>=4.2.5

# 系统信息
platform>=1.0.8

# 文件操作
pathlib2>=2.3.7

# -----------------------------------------------------------------------------
# 时间和调度
# -----------------------------------------------------------------------------
# 任务调度
schedule>=1.2.0
APScheduler>=3.10.1

# 时间处理
arrow>=1.2.3
pendulum>=2.1.2

# -----------------------------------------------------------------------------
# 数据验证和模型
# -----------------------------------------------------------------------------
# 数据验证
pydantic>=2.1.1

# 类型提示
typing-extensions>=4.7.1

# 数据类
dataclasses>=0.8

# -----------------------------------------------------------------------------
# 图像和媒体处理（可选）
# -----------------------------------------------------------------------------
# 图像处理
Pillow>=10.0.0

# 图像下载和处理
requests-cache>=1.1.0

# -----------------------------------------------------------------------------
# 容器化和部署（可选）
# -----------------------------------------------------------------------------
# Docker支持
docker>=6.1.3

# 环境检测
platform>=1.0.8

# -----------------------------------------------------------------------------
# 其他实用工具
# -----------------------------------------------------------------------------
# 进度条
tqdm>=4.65.0

# 颜色输出
colorama>=0.4.6

# 富文本输出
rich>=13.4.2

# 命令行参数解析
click>=8.1.6

# 配置文件解析
PyYAML>=6.0.1

# 随机数据生成
Faker>=19.3.0